Added Command Query Demo

+++ CommandQuery.Api/WeatherForecast.cs
@@ -1,12 +1,18 @@
+using System;
+
 namespace CommandQuery.Api;
 
 public class WeatherForecast
 {
     public DateTime Date { get; set; }
 
-    public int TemperatureC { get; set; }
+    public float TemperatureC { get; set; }
 
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+    public float TemperatureF => 32 + (TemperatureC / 0.5556f);
+
+#nullable enable
 
     public string? Summary { get; set; }
+
+#nullable disable
 }
